GantBackend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
protobuf==3.20.3
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
mysqlclient==2.1.1